alterei o tipo da variavel

+++ PriceTravilling.py
@@ -1,4 +1,4 @@
-km = int(input('Informe a distância do seu destino: '))
+km = float(input('Informe a distância do seu destino: '))
 if km <= 200:
     passagem = km * 0.50
     print('Sua passagem é no valor de R$ {:.2f}'.format(passagem))
